+++ view/api_v4_get_product_detail.py
@@ -1,42 +1,102 @@
+from view.utils import timer
+
 import os
 import json
-import time
+import logging
 import asyncio
 import datetime
 
 import aiohttp
 import pandas as pd
 
+from pydantic import BaseModel
 
-class CrawlerSearchItems:
+logger = logging.getLogger()
+
+
+class ItemParams(BaseModel):
+    itemid: str
+    shopid: int
+    name: str
+    currency: str
+    stock: int
+    status: int
+    ctime: int
+    t_ctime: str
+    sold: int
+    historical_sold: int
+    liked_count: int
+    brand: str
+    cmt_count: int
+    item_status: str
+    price: int
+    price_min: int
+    price_max: int
+    price_before_discount: int
+    show_discount: int
+    raw_discount: int
+    tier_variations_option: str
+    rating_star_avg: int
+    rating_star_1: int
+    rating_star_2: int
+    rating_star_3: int
+    rating_star_4: int
+    rating_star_5: int
+    item_type: int
+    is_adult: bool
+    has_lowest_price_guarantee: bool
+    is_official_shop: bool
+    is_cc_installment_payment_eligible: bool
+    is_non_cc_installment_payment_eligible: bool
+    is_preferred_plus_seller: bool
+    is_mart: bool
+    is_on_flash_sale: bool
+    is_service_by_shopee: bool
+    shopee_verified: bool
+    show_official_shop_label: bool
+    show_shopee_verified_label: bool
+    show_official_shop_label_in_title: bool
+    show_free_shipping: bool
+
+    class Config:
+        allow_extra = False
+
+
+class CrawlerProductDetail:
     def __init__(self):
         self.basepath = os.path.abspath(os.path.dirname(__file__))
 
-        self.search_item_api = "https://shopee.tw/api/v4/shop/search_items?filter_sold_out=1&limit=30&offset=1&order=desc&shopid=5547415&sort_by=pop&use_case=4"
-        self.item = {}
+        self.search_item_api = "https://shopee.tw/api/v4/shop/search_items"
+        self.items_info = []
 
-    def __call__(self, input_shop_ids):
+    @timer
+    def __call__(self, shop_detail):
         async def parser_shop_html(html):
-            shop = json.loads(html)
+            info = json.loads(html)
 
-            dateArray = datetime.datetime.utcfromtimestamp(shop["data"]["ctime"])
-            transfor_time = dateArray.strftime("%Y-%m-%d %H:%M:%S")
-            self.shop_detail_dict["shop_ctime"].append(transfor_time)
-            self.shop_detail_dict["shopid"].append(shop["data"]["shopid"])
-            self.shop_detail_dict["shop_name"].append(shop["data"]["name"])
-            self.shop_detail_dict["shop_country"].append(shop["data"]["country"])
-            self.shop_detail_dict["shop_item_count"].append(shop["data"]["item_count"])
-            self.shop_detail_dict["shop_place"].append(shop["data"]["place"])
-            self.shop_detail_dict["shop_rating_star"].append(
-                shop["data"]["rating_star"]
-            )
-            self.shop_detail_dict["shop_rating_bad"].append(shop["data"]["rating_bad"])
-            self.shop_detail_dict["shop_rating_normal"].append(
-                shop["data"]["rating_normal"]
-            )
-            self.shop_detail_dict["shop_rating_good"].append(
-                shop["data"]["rating_good"]
-            )
+            if info["total_count"] != 0:
+                for item in info["items"]:
+                    item = item["item_basic"]
+
+                    dateArray = datetime.datetime.utcfromtimestamp(item["ctime"])
+                    transfor_time = dateArray.strftime("%Y-%m-%d %H:%M:%S")
+
+                    item_info = ItemParams(
+                        **item,
+                        t_ctime=transfor_time,
+                        rating_star_avg=item["item_rating"]["rating_star"],
+                        rating_star_1=item["item_rating"]["rating_count"][1],
+                        rating_star_2=item["item_rating"]["rating_count"][2],
+                        rating_star_3=item["item_rating"]["rating_count"][3],
+                        rating_star_4=item["item_rating"]["rating_count"][4],
+                        rating_star_5=item["item_rating"]["rating_count"][5],
+                        tier_variations_option=",".join(
+                            item["tier_variations"][0]["options"]
+                        )
+                        if item.get("tier_variations")
+                        else "",
+                    )
+                    self.items_info.append(item_info.dict())
 
         async def get_item_detail(client, query_url):
             try:
@@ -45,9 +105,10 @@ class CrawlerSearchItems:
                     assert response.status == 200
                     await parser_shop_html(html)
             except Exception as e:
+                logger.warning(f"Exception: {e}")
                 print("---Exception---:", e)
 
-        async def main(crawler_shop_urls):
+        async def main(crawler_itme_urls):
             headers = {
                 "User-Agent": "facebookexternalhit/1.1 (+http://www.facebook.com/externalhit_uatext.php)",
                 "referer": "https://shopee.tw/",
@@ -59,45 +120,59 @@ class CrawlerSearchItems:
             ) as client:
                 tasks = [
                     get_item_detail(client, query_url)
-                    for query_url in crawler_shop_urls
+                    for query_url in crawler_itme_urls
                 ]
                 await asyncio.gather(*tasks)
 
-        crawler_shop_urls = []
-        for id in range(len(input_shop_ids)):
-            crawler_shop_urls.append(self.shop_detail_api + str(input_shop_ids[id]))
-        asyncio.run(main(crawler_shop_urls))
+        crawler_itme_urls = []
 
-        df = pd.DataFrame(self.shop_detail_dict)
-        df.to_csv(self.basepath + "/csv/shop_detail.csv", index=False)
-        return df
+        df_header = pd.DataFrame(
+            columns=[field.name for field in ItemParams.__fields__.values()]
+        )
+        df_header.to_csv(self.basepath + "/csv/pdp_detail.csv", index=False)
+
+        for row in shop_detail.itertuples():
+
+            shop_id = row.shopid
+            shop_product_count = row.item_count
+            num = 0
+            while num < shop_product_count:
+                crawler_itme_urls.append(
+                    f"{self.search_item_api}?offset={str(num)}&limit=100&order=desc&filter_sold_out=3&use_case=1&sort_by=sales&order=sales&shopid={shop_id}"
+                )
+                num += 100
+            asyncio.run(main(crawler_itme_urls))
+
+            df = pd.DataFrame(self.items_info)
+            df.to_csv(
+                self.basepath + "/csv/pdp_detail.csv",
+                index=False,
+                mode="a",
+                header=False,
+            )
+        total_df = pd.read_csv(self.basepath + "/csv/pdp_detail.csv")
+        return total_df
 
 
 if __name__ == "__main__":
-    # // api example
-    # https://shopee.tw/api/v4/shop/search_items?filter_sold_out=1&limit=30&offset=1&order=desc&shopid=5547415&sort_by=pop&use_case=4
+    """
+    # api example
+    # https://shopee.tw/api/v4/shop/search_items?filter_sold_out=1&limit=100&offset=1&order=desc&shopid=5547415&sort_by=pop&use_case=1
 
-    time_start = time.time()
+    params use_case:
+    1: Top Product
+    2: ?
+    3: ?
+    4: Sold out items
 
-    input_shop_ids = [
-        5547415,
-        # 22189057,
-        # 1517097,
-        # 3323966,
-        # 1971812,
-        # 8016627,
-        # 80078149,
-        # 7314701,
-        # 151143321,
-        # 47924061,
-        # 29951329,
-        # 9532352,
-        # 15659558,
-        # 31945247,
-    ]
+    params filter_sold_out:
+    1: = sold_out
+    2: != sold_out
+    3: both
 
-    do = CrawlerSearchItems()
-    result = do(input_shop_ids)
+    """
 
-    print(result)
-    print(time.time() - time_start)
+    basepath = os.path.abspath(os.path.dirname(__file__))
+    shop_detail = pd.read_csv(basepath + "/csv/shop_detail.csv")
+    crawler_product_detail = CrawlerProductDetail()
+    result_product_detail = crawler_product_detail(shop_detail)