+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Feat
 
+- **config.py**: add logging function
 - **shop_detail**: upgrade to api v4
 - **utils.py**: add timer function to count the running time
 - **main.py**: add ip pool function